refactor: remove drawer example as it does not add to overall example
