+++ examples/reactnavigation/jest-setup.js
@@ -1,14 +1,2 @@
-// require('react-native-reanimated/lib/reanimated2/jestUtils').setUpTests();
-
-// jest.mock('react-native-reanimated', () => {
-//   const Reanimated = require('react-native-reanimated/mock');
-
-//   // The mock for `call` immediately calls the callback which is incorrect
-//   // So we override it with a no-op
-//   Reanimated.default.call = () => {};
-
-//   return Reanimated;
-// });
-
 // Silence the warning: Animated: `useNativeDriver` is not supported because the native animated module is missing
 jest.mock('react-native/Libraries/Animated/NativeAnimatedHelper');
